requirements.txt
# Core Dependencies
gradio>=4.0.0
python-dotenv>=1.0.0

# Azure SDK
azure-cognitiveservices-speech>=1.35.0
azure-search-documents>=11.4.0
azure-core>=1.29.0
azure-storage-blob>=12.19.0

# OpenAI
openai>=1.10.0

# Document Processing
PyPDF2>=3.0.0
python-docx>=1.1.0
python-pptx>=0.6.23

# PDF Generation
reportlab>=4.0.0

# Audio Processing
soundfile>=0.12.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0
Pillow>=10.0.0

# Optional: Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
